UI: Add Formsy input for creation mode selection step
